refactor(fetcher): move hackernews into family directory with shared client
Promotes `hackernews_top.rs` into `src/fetcher/hackernews/` so siblings
(`hackernews_user`, `hackernews_user_submissions`, `hackernews_user_comments`)
can share the HTTP client setup, base URL constants, and the OSC 8 / HN
user-page link helpers without each file open-coding `OnceLock<Client>` and
the Firebase host. Mirrors the layout used by the `github_*` family.

No behaviour change: only `hackernews_top` ships in this commit; the
sibling fetchers land in the next commit.
